Fixes Segfault on fio with multiple jobfiles
When running fio with multiple jobfiles the nvmed_fio driver segfaults.
The reason is because the fio_nvmed_io_u_init is called for the second
jobfile and the nvmed.nvmed is NULL resulting in a null pointer
dereference when calling nvmed_get_buffer(nvmed.nvmed, ...).

The reason is because the nvmed dev file was closed by fio_nvmed_cleanup
after the first execution. As the dev file should be open throughout the
execution of the fio, moved the nvmed open and close to the open/close
functions.