Adds error checking to mmap error in lib
The default mlock allowance on my system was 64KB. NVMeDirect requested
for 64 * 4KB locked pages, resulting in a mmap fail.
However the library does not handle a mmap fail, and thus a segfault
occurs when trying to clear to mmaped region.

This commit adds mmap fail handling, and also suggest checking the mlock
limit.